FD-3157: antd style import problem is fixed
