--- components/PersonalFinance/TransactionPage.css
@@ -1,36 +1,0 @@
-/* .ant-picker-cell-in-view.ant-picker-cell-selected .ant-picker-cell-inner, .ant-picker-cell-in-view.ant-picker-cell-range-start .ant-picker-cell-inner, .ant-picker-cell-in-view.ant-picker-cell-range-end .ant-picker-cell-inner{
-    background-color:#4A56E2;
-  } */
-  /* .ant-picker-cell-in-view.ant-picker-cell-today .ant-picker-cell-inner::before {
-  border: none;
-  } */
-  .ant-picker-header {
-    color: #4A56E2 !important;
-  }
-  .ant-picker-cell-in-view.ant-picker-cell-selected .ant-picker-cell-inner {
-    background-color: #E0EAFF !important ; 
-  }
-  .ant-picker-cell-in-view.ant-picker-cell-selected .ant-picker-cell-inner .selected-date {
-    font-weight: bold;
-    font-size: 17px !important;
-  }
-  .ant-picker-cell-in-view.ant-picker-cell-selected .ant-picker-cell-inner .selected-amount { 
-    font-weight: bold;
-  }
- 
-  .ant-picker-cell-in-view.ant-picker-cell-today .ant-picker-cell-inner::before {
-    border: none !important;
-  }
-  .ant-picker-cell .ant-picker-cell-inner {
-    border-radius: 0.5rem !important;
-    color: #4A56E2 !important;
-    height:3.3rem !important;
-  }
-  .custom-picker input {
-    visibility: hidden;
-    height: 0;
-    padding: 0;
-    width: 0;
-    position: absolute;
-  }
-
