+++ pages/_app.js
@@ -1,7 +1,6 @@
 import { ToastNotification } from '../components/ToastNotification';
 import { AccountVerificationFormProvider } from '../components/AccountVerificationForm';
 import '../styles.css';
-import '../components/PersonalFinance/TransactionPage.css'
 
 export default function MyApp({ Component, pageProps }) {
   return (
